Initialize "compare" benchmarks
